+++ ngmix/priors/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .priors import *
+from .random import make_rng, srandu