+++ CHANGES.md
@@ -9,6 +9,8 @@
 - fixed large rounding errors in some shape conversions in `ngmix.shape`
 - fixed bug in `ngmix.shape.g1g2_to_eta1eta2` in handling numpy array inputs
 - fixed bug in `ngmix.shape.e1e2_to_eta1eta2` in handling numpy array inputs
+- fixed bug in `ngmix.priors.GPriorBA.get_fdiff` in handling array inputs
+- fixed a bug in `ngmix.priors.TwoSidedErf.get_fdiff` in handling array inputs
 
 ### new features
 
@@ -17,6 +19,7 @@
   * `ngmix.jacobian`
   * `ngmix.moments`
   * `ngmix.observation`
+  * `ngmix.priors`
   * `ngmix.gexceptions`
 - introduced `NGmixBaseException` as parent class for all ngmix-specific exceptions
 
@@ -27,6 +30,8 @@
   along with the code used to test them, have been removed from the classes in
   `ngmix.priors`
 - removed travis-ci in favor of GitHub Actions
+- the optional `rng` keyword for `ngmix.priors.srandu` has been removed in favor
+  of a required keyword
 
 
 ## v1.3.8